pyproject.toml
[project]
name = "resume-screening"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langextract>=1.1.1",
    "ollama>=0.6.1",
    "pymupdf>=1.27.2",
    "python-dotenv>=1.2.2",
    "unstructured-inference>=1.5.2",
    "unstructured[docx,pdf]>=0.21.5",
]
